Rename/reorganize some paths.
