+++ src/index.ts
@@ -25,7 +25,9 @@ import { Debugger } from './debugger';
 // import { DebuggerSidebar } from './sidebar';
 
 import { IDebugger, IDebuggerSidebar } from './tokens';
-import { DebuggerNotebookTracker } from './notebookTracker';
+
+import { DebuggerNotebookTracker } from './handlers/notebook';
+
 import { BreakpointsService, SessionTypes } from './breakpointsService';
 
 import { DebuggerConsoleTracker } from './consoleTracker';
